Make if statements work and revamp them. Read more:
New type: eqexpression - this is just to prevent shit like and and
or from getting caught in the math parser, and to help perform
error handling. Because of this, if statement expressions don't
have to be wrapped in quotes. Also actually made if statements
work and did some checks for calling commands with no args.