Better error handling, and huge bugfix (see below)
This commit has been in the works for about 4 days now. The bug had to do with the language not allowing nested loops and if statements. That is fixed now, and it was fixed by adding a stack of strings, one for each block. Also, much better error handling ensues.